Backend/weather
Completed the weather function and the UI.
