+++ public/css/spot.css
@@ -502,21 +502,84 @@ button:focus {
     cursor: pointer;
 }
 
-.weather-icon { 
-    display: flex; align-items: center; /* アイコンとテキストを縦に中央揃え */ 
+.info-container { 
+    display: flex; 
+    justify-content: space-between;
+    align-items: stretch; 
+} 
+
+.map {
+     flex: 2; /* 地図の幅を広くする */ 
+     margin: 10px; display: flex; 
+     flex-direction: column; 
+     justify-content: space-between;
 }
 
-.weather-icon img {
-    width: 70px; /* アイコンの幅を50pxに設定 */
-    height: 70px; /* アイコンの高さを50pxに設定 */
+.weather { 
+    flex: 1; 
+    margin: 10px; 
+    display: flex; 
+    flex-direction: column;
+    justify-content: space-between; /* 高さを揃えるために追加 */ 
+    padding: 20px; 
+    text-align: center;
 }
 
-.weather-icon .temperature { 
-    font-size: 50px; /* 温度の文字サイズを画像と同じくらい大きくする */ 
-    font-weight: bold; /* 太字にして目立たせる */ 
-    margin-left: 20px; /* アイコンと温度の間にスペースを追加する */ 
+.weather h3 {
+    margin-bottom: 10px;
 }
 
 .weather-condition {
-    font-size: 18px; /* 文字サイズを調整 */
+    font-size: 24px;
+    margin-bottom: 5px;
+}
+
+.weather-icon-container { 
+    display: flex; 
+    align-items: center; 
+    justify-content: center; 
+    margin-bottom: 10px; /* マージンを追加して下部スペースを調整 */
+}
+
+.weather-icon img {
+    background-color: #add8e6; /* 背景色を水色に設定 */
+    padding: 5px;
+    width: 100px;
+    height: 100px;
+    margin-right: 15px; /* アイコンと気温の間にスペースを追加 */
+}
+
+.temperature {
+    font-size: 48px
+}
+
+.weather-info { 
+    display: flex; 
+    flex-direction: column; 
+    align-items: flex-start; 
+} 
+
+.weather-info p {
+     margin: 0; 
+    }
+
+.weather-details {
+    font-size: 20px;
+    margin-top: 15px; /* マージンを調整 */
+}
+
+.weather-details p {
+    margin: 5px 0;
+}
+
+.large-number { 
+    font-size: 25px; /* 特定の部分のフォントサイズを大きく設定 */ 
+    font-weight: bold; /* 強調表示 */ 
+}
+
+.card3 { 
+    flex: 1; /* 高さを揃えるために追加 */ 
+    display: flex; 
+    flex-direction: column; 
+    justify-content: space-between; 
 }
