+++ routes/web.php
@@ -213,8 +213,6 @@ Route::get('/select-post-form', function () {
 })->name('select-post-form');
 
 
-Route::get('/spot/{spot_id}', [WeatherController::class, 'show']);
-
 // Authentication Routes
 
 Auth::routes();
